2059: Moved distribution chart into own class
